+++ responder/__init__.py
@@ -7,12 +7,13 @@ including the API, Request, and Response classes.
 
 from . import ext
 from .__version__ import __version__
-from .core import API, Request, Response
+from .core import API, Request, Response, abort
 
 __all__ = [
     "API",
     "Request",
     "Response",
     "__version__",
+    "abort",
     "ext",
 ]