+++ responder/core.py
@@ -1,8 +1,9 @@
-from .api import API
+from .api import API, abort
 from .models import Request, Response
 
 __all__ = [
     "API",
     "Request",
     "Response",
+    "abort",
 ]